Created a file to store custom types, custom data types that are generic enough to be used by multiple components will be stored here
